@@ -2,7 +2,6 @@ package com.digero.maestro.abc;
 
 import java.util.AbstractCollection;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -47,7 +46,7 @@ public class LotroStudentFXInfo implements Comparable<LotroStudentFXInfo>
 			}
 		}
 
-		ALL_FX = Collections.unmodifiableList(new ArrayList<>(new AbstractCollection<LotroStudentFXInfo>() {
+		ALL_FX = List.copyOf(new AbstractCollection<>() {
 			@Override
 			public Iterator<LotroStudentFXInfo> iterator() {
 				return new FXInfoIterator();
@@ -57,7 +56,7 @@ public class LotroStudentFXInfo implements Comparable<LotroStudentFXInfo>
 			public int size() {
 				return byId.size();
 			}
-		}));
+		});
 	}
 
 //	private static final Comparator<Note> noteComparator = new Comparator<Note>() {
